refactor: 과거 예약 Custom 예외 도메인으로 이동
